requirements.txt
aioitertools==0.12.0
aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
arize-phoenix==11.21.0
arize-phoenix-client==1.15.0
arize-phoenix-evals==0.26.1
arize-phoenix-otel==0.12.1
attrs==25.3.0
Authlib==1.6.1
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cryptography==45.0.5
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
email_validator==2.2.0
fastapi==0.116.1
google-adk==1.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.105.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.35.1
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.27.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphql-core==3.2.6
graphviz==0.21
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
MarkupSafe==3.0.2
mcp==1.12.2
numpy==2.3.2
openai==1.99.3
openinference-instrumentation==0.1.37
openinference-instrumentation-google-adk==0.1.3
openinference-instrumentation-openai==0.1.30
openinference-semantic-conventions==0.1.21
opentelemetry-api==1.36.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
prometheus_client==0.22.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing==3.2.3
pystache==0.6.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlean.py==3.49.1
sse-starlette==2.4.1
starlette==0.47.2
strawberry-graphql==0.270.1
tenacity==8.5.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
wrapt==1.17.2
zipp==3.23.0
